+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where a small amount of extra etcd or postgres traffic was
+generated on keepalive failure.
+
 ## [6.2.0] - 2020-12-17
 
 ### Added