+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 - Fixed a bug where a small amount of extra etcd or postgres traffic was
 generated on keepalive failure.
+- Allow agents labels & annotations to be configured via configuration flags.
 
 ## [6.2.0] - 2020-12-17