+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+## [6.2.1] - 2021-01-07
+
 ### Fixed
 - The expire field of silenced entries represents the configured expiration, in
 seconds, not the remaining duration.